# nightclub-app/spec/current/openapi.yaml
openapi: 3.1.1
info:
  title: Nightclub v4 R1 Core API Design
  version: 4.0.0
  description: 開発用コア契約。未実装。ユーザー確定：名前検索中心、入口も通常承認、入口共用端末。R1全機能の詳細化状況はapi/coverage.mdを参照。実サービスURL・認証基盤・PSPは未確定。
servers:
- url: https://api.example.invalid/v1
  description: 非実在の例示URL。実接続先ではない。
security: &id001
- PersonalSession: []
- DeviceSession: []
  OperatorSession: []
paths:
...
  /stores/{storeId}/events/{eventId}/snapshot:
    get:
      operationId: getSnapshot
      tags:
      - Sync
      summary: 一貫した受付スナップショット
      description: '


        権限：sync.read。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: sync.read
      x-requirement-ids:
      - V3-SYNC-03
      - V4-DEV-03
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: eventId
        in: path
        required: true
        schema: *id003
      - name: page_token
        in: query
        required: false
        schema: *id002
        description: 同じスナップショットの続き
      - $ref: '#/components/parameters/OperatorContext'
  /stores/{storeId}/events/{eventId}/changes:
    get:
      operationId: getChanges
      tags:
      - Sync
      summary: カーソル以降の権限内変更
      description: '


        権限：sync.read。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: sync.read
      x-requirement-ids:
      - V3-SYNC-02
      - V4-DEV-03
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangesResult'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: eventId
        in: path
        required: true
        schema: *id003
      - name: cursor
        in: query
        required: true
        schema: *id002
        description: スナップショットで取得した署名付きカーソル
      - name: limit
        in: query
        required: false
        schema: *id004
        description: 最大100
      - $ref: '#/components/parameters/OperatorContext'
  /stores/{storeId}/events/{eventId}/sync-ack:
    post:
      operationId: ackSync
      tags:
      - Sync
      summary: 入口の画面へ適用した位置を記録
      description: '


        権限：sync.ack。入口モードでは有効な端末・担当者・当日役割を検証。主体IDをbodyで偽装できない。'
      security: *id001
      x-permission: sync.ack
      x-requirement-ids:
      - V3-SYNC-04
      responses:
        '200':
          description: 成功（外部処理待ちは状態で明示）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ack'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: storeId
        in: path
        required: true
        schema: *id003
      - name: eventId
        in: path
        required: true
        schema: *id003
      - $ref: '#/components/parameters/CsrfToken'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/OperatorContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncAck'
...
components:
  securitySchemes:
    PersonalSession:
      type: apiKey
      in: cookie
      name: __Host-nc_personal
    DeviceSession:
      type: apiKey
      in: cookie
      name: __Host-nc_device
    OperatorSession:
      type: apiKey
      in: cookie
      name: __Host-nc_operator
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 160
      description: 同じ業務操作は同じキー。異なるpayloadへの流用は409。金銭は保存済み参照も照合。
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: *id002
      description: cookie認証の状態変更に必要。Origin/Fetch Metadataも検証。
    OperatorContext:
      name: X-Operator-Context
      in: header
      required: false
      schema: *id003
      description: 入口モードでは必須。画面描画時のoperator_session_id。認証手段ではない。cookie主体とサーバーactive sessionに一致しなければ409 OPERATOR_CHANGED。個人モードでは不要。
  schemas:
    Problem:
      type: object
      properties:
        type: *id002
        title: *id002
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail: *id002
        code: *id002
        trace_id: *id003
        current_version:
          type:
          - integer
          - 'null'
        retryable:
          type: boolean
        field_errors:
          type: array
          items:
            type: object
            properties:
              path: *id002
              code: *id002
              message: *id002
            required:
            - path
            - code
            - message
            additionalProperties: false
      required:
      - type
      - title
      - status
      - code
      - trace_id
      - retryable
      additionalProperties: false
    Ack:
      type: object
      properties:
        accepted:
          type: boolean
        trace_id: *id003
      required:
      - accepted
      - trace_id
      additionalProperties: false
...
    SegmentSummary:
      type: object
      properties:
        id: *id003
        version: *id005
        status:
          type: string
          enum:
          - PENDING
          - AUTHORIZED
          - REJECTED
          - RETURNED
          - REVOKED
          - EXPIRED
        requested_count: *id005
        authorized_count:
          type: integer
          minimum: 0
        first_entered_count:
          type: integer
          minimum: 0
        remaining_count:
          type: integer
          minimum: 0
        unit_amount_minor: &id008
          type: integer
          minimum: 0
          maximum: 9007199254740991
        currency:
          type: string
          pattern: ^[A-Z]{3}$
        authorization_method:
          type:
          - string
          - 'null'
          enum:
          - STANDARD
          - TRUSTED_ACTOR
          - CUSTOMER_PERMIT
          - MANUAL
          - null
        requires_identity_check:
          type: boolean
        entry_until: *id007
      required:
      - id
      - version
      - status
      - requested_count
      - authorized_count
      - first_entered_count
      - remaining_count
      - unit_amount_minor
      - currency
      - authorization_method
      - requires_identity_check
      - entry_until
      additionalProperties: false
    VisitSummary:
      type: object
      properties:
        id: *id003
        version: *id005
        reception_name: *id002
        customer_id: *id006
        referrer_membership_id: *id006
        referrer_name:
          type:
          - string
          - 'null'
        arrival_status: *id002
        status: *id002
        segments:
          type: array
          items:
            $ref: '#/components/schemas/SegmentSummary'
        entry_device_sync:
          type: object
          properties:
            active_devices:
              type: integer
              minimum: 0
            synced_devices:
              type: integer
              minimum: 0
          required:
          - active_devices
          - synced_devices
          additionalProperties: false
      required:
      - id
      - version
      - reception_name
      - customer_id
      - referrer_membership_id
      - referrer_name
      - arrival_status
      - status
      - segments
      - entry_device_sync
      additionalProperties: false
...
    SnapshotResult:
      type: object
      properties:
        snapshot_token: *id002
        cursor: *id002
        expires_at: *id007
        visits:
          type: array
          items:
            $ref: '#/components/schemas/VisitSummary'
        has_more:
          type: boolean
        next_page_token:
          type:
          - string
          - 'null'
      required:
      - snapshot_token
      - cursor
      - expires_at
      - visits
      - has_more
      - next_page_token
      additionalProperties: false
    ChangeEvent:
      type: object
      properties:
        event_id: *id003
        stream_seq:
          type: integer
          minimum: 1
        event_type: *id002
        aggregate_type: *id002
        aggregate_id: *id003
        aggregate_version: *id005
        action:
          type: string
          enum:
          - UPSERT
          - INVALIDATE
          - REMOVE
        data:
          type:
          - object
          - 'null'
          description: 権限に応じたVisitSummaryまたは最小無効化情報。raw outboxの転送は禁止。
      required:
      - event_id
      - stream_seq
      - event_type
      - aggregate_type
      - aggregate_id
      - aggregate_version
      - action
      - data
      additionalProperties: false
    ChangesResult:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/ChangeEvent'
        next_cursor: *id002
        reset_required:
          type: boolean
        has_more:
          type: boolean
      required:
      - events
      - next_cursor
      - reset_required
      - has_more
      additionalProperties: false
    SyncAck:
      type: object
      properties:
        snapshot_token: *id002
        cursor: *id002
        visibility:
          type: string
          enum:
          - FOREGROUND
          - BACKGROUND
      required:
      - snapshot_token
      - cursor
      - visibility
      additionalProperties: false
...
  responses:
    Problem:
      description: 構造化エラー。detailsで個人情報・内部SQLを漏らさない。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'